requirements.txt
asgiref==3.2.10
Django==3.1.1
django-bootstrap3==14.1.0
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
django-bootstrap3==14.1.0
